+++ frontend/lib/screens/user/edit_profile.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import '../provider/skills_edit_widget.dart';
-import '../provider/achievements_edit_widget.dart';
-import '../provider/about_Edit_widget.dart';
+import '../../provider/skills_edit_widget.dart';
+import '../../provider/achievements_edit_widget.dart';
+import '../../provider/about_Edit_widget.dart';
 import 'dart:convert';
 
 class EditProfileScreen extends ConsumerWidget {
